refactor(nms): Separate the GatewayPoolsContext
